+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 0.3.0
+- [BREAKING CHANGE] Separate configuration file into two, credentials and cluster settings files
 - Added external method for setting chunk size (SetChunkSize)
+- Added getJobList function to check the status of user's jobs